% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-docs.R
\docType{data}
\name{survey_ghl}
\alias{survey_ghl}
\title{Greenland halibut measurements from IMR surveys}
\format{
A dataframe
}
\source{
Institute of Marine Research (\url{https://www.imr.no})
}
\usage{
data(survey_ghl)
}
\description{
Greenland halibut measurements from IMR surveys
}
\details{
Contains length, weight, age, sex and maturity measurements of Greenland halibut acquired on various surveys.
}
\keyword{datasets}
