% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_fishplots.R
\name{theme_fishplots}
\alias{theme_fishplots}
\title{A ggplot2 theme for the ggFishPlots package}
\usage{
theme_fishplots(...)
}
\arguments{
\item{...}{additional arguments passed to \code{\link[ggplot2]{ggtheme}}.}
}
\value{
A ggplot2 theme layer.
}
\description{
A ggplot2 theme for the ggFishPlots package
}
\concept{customize shapefiles}
