% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{unlogit}
\alias{unlogit}
\title{Back-transform predictor variables from a logit model}
\usage{
unlogit(p, model)
}
\arguments{
\item{p}{probability threshold to unlog from \code{model}}

\item{model}{a \code{glm} model object}
}
\value{
A data frame
}
\description{
Back-transform predictor variables from a logit model
}
\keyword{internal}
