% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lw.R
\name{plot_lw}
\alias{plot_lw}
\title{Plot length-weight relationships}
\usage{
plot_lw(
  dt,
  length = "length",
  weight = "weight",
  sex = "sex",
  female.sex = "F",
  male.sex = "M",
  length.unit = "cm",
  weight.unit = "kg",
  split.by.sex = FALSE,
  filter.exp = NULL,
  xlab = "Total length",
  ylab = "Weight",
  use.nls = FALSE,
  log.axes = FALSE,
  base_size = 8,
  legend.position = "bottom"
)
}
\arguments{
\item{dt}{A data.frame, tibble or data.table}

\item{length}{Character argument giving the name of the length column in \code{dt}}

\item{weight}{Character argument giving the name of the age column in \code{dt}}

\item{sex}{Character argument giving the name of the sex column in \code{dt}. Ignored if \code{split.by.sex == FALSE}.}

\item{female.sex}{A character denoting female sex in the \code{sex} column of \code{dt}}

\item{male.sex}{A character denoting male sex in the \code{sex} column of \code{dt}}

\item{length.unit}{A character argument giving the unit of \code{length}. Will be used in the labels of the figure.}

\item{weight.unit}{Character argument giving the unit of \code{weight}. Will be used in the labels of the figure.}

\item{split.by.sex}{Logical indicating whether the result should be split by sex.}

\item{filter.exp}{Expression to filter data. E.g. 'sampling_type == "ENS"'}

\item{xlab}{Character giving the x-axis label without unit}

\item{ylab}{Character giving the x-axis label without unit.}

\item{use.nls}{Logical indicating whether the parameters should be calculated using the nonlinear least squares (\code{nls; TRUE}) method over the log-log transformed linear model (\code{lm; FALSE}) method.}

\item{log.axes}{Logical indicating whether logarithmic axes should be used instead of cartesian ones.}

\item{base_size}{Base size parameter for ggplot. See \link[ggplot2]{ggtheme}.}

\item{legend.position}{Position of the ggplot legend as a character. See \link[ggplot2]{ggtheme}.}
}
\value{
A ggplot together with the a and b parameters.
}
\description{
Plot length-weight relationships
}
\examples{
data(survey_ghl)

# Simple plot
plot_lw(survey_ghl, length = "length", weight = "weight")

# Split by sex
plot_lw(survey_ghl, split.by.sex = TRUE)$plot
}
\author{
Mikko Vihtakari // Institute of Marine Research.
}
