% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_maturity.R
\name{plot_maturity}
\alias{plot_maturity}
\title{Plot maturity ogive}
\usage{
plot_maturity(
  dt,
  length = "length",
  maturity = "maturity",
  sex = "sex",
  split.by.sex = FALSE,
  female.sex = "F",
  male.sex = "M",
  length.unit = "cm",
  length.bin.width = 2,
  force.zero.group.length = NA,
  force.zero.group.strength = 10,
  force.zero.group.cv = 0,
  xlab = "Total length",
  base_size = 8,
  legend.position = "bottom"
)
}
\arguments{
\item{dt}{A data.frame, tibble or data.table}

\item{length}{Character argument giving the name of the length (or age) column in \code{dt}}

\item{maturity}{Character argument giving the name of the maturity column in \code{dt}. Should be either logical (\code{TRUE == mature, FALSE == immature}) or integer (\code{1 == mature, 0 == immature}).}

\item{sex}{Character argument giving the name of the sex column in \code{dt}. Ignored if \code{split.by.sex == FALSE}.}

\item{split.by.sex}{Logical indicating whether the result should be split by sex.}

\item{female.sex}{A character denoting female sex in the \code{sex} column of \code{dt}}

\item{male.sex}{A character denoting male sex in the \code{sex} column of \code{dt}}

\item{length.unit}{A character argument giving the unit of \code{length}. Will be used in the labels of the figure.}

\item{length.bin.width}{Numeric specifying the increment (delta length) by which length data should be binned to calculate maturity proportions. Use \code{NULL} to remove from the plot.}

\item{force.zero.group.length}{Numeric indicating the length to which 0-group (all immatures) should be forced. Use \code{NA} ignore the forcing.
#' @param force.zero.group.cv Numeric indicating the coefficient of variation for the forced 0-group (all immature) length. Resulting lengths will be randomly generated.}

\item{force.zero.group.strength}{Numeric indicating how many percent of total fish should be added to the specified \code{force.zero.group.length}.}

\item{xlab}{Character giving the x-axis label without unit}

\item{base_size}{Base size parameter for ggplot. See \link[ggplot2]{ggtheme}.}

\item{legend.position}{Position of the ggplot legend as a character. See \link[ggplot2]{ggtheme}.}

\item{...}{Additional arguments passed to \link[ggridges]{geom_density_ridges}.}
}
\value{
Returns a ggplot2 or tibble depending on the \code{plot} argument showing the maturity ogives.
}
\description{
Plots an estimate of length or age at 50\% mature for a dataset
}
\details{
Depends on the tidyverse and ggridges packages. The dplyr and ggplot2 packages must be loaded into the workspace.
}
\examples{
# Simple L50 plot
data(survey_ghl)
plot_maturity(survey_ghl, length = "length", maturity = "maturity")

# A50 plot, split by sex
plot_maturity(survey_ghl, length = "age", length.unit = "years",
xlab = "Age", length.bin.width = 1, split.by.sex = TRUE)$plot
}
\author{
Mikko Vihtakari // Institute of Marine Research.
}
