% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_maturity.R
\name{plot_maturity}
\alias{plot_maturity}
\title{Plot maturity ogive}
\usage{
plot_maturity(
  dt,
  length = "length",
  maturity = "maturity",
  sex = "sex",
  female.sex = "F",
  male.sex = "M",
  length.unit = "cm",
  length.bin.width = 2,
  split.by.sex = FALSE,
  filter.exp = NULL,
  xlab = "Total length",
  base_size = 8
)
}
\arguments{
\item{dt}{A data.frame, tibble or data.table}

\item{length}{Character argument giving the name of the length (or age) column in \code{dt}}

\item{maturity}{Character argument giving the name of the maturity column in \code{dt}. Should be either logical (\code{TRUE == mature, FALSE == immature}) or integer (\code{1 == mature, 0 == immature}).}

\item{sex}{Character argument giving the name of the sex column in \code{dt}. Ignored if \code{split.by.sex == FALSE}.}

\item{female.sex}{A character denoting female sex in the \code{sex} column of \code{dt}}

\item{male.sex}{A character denoting male sex in the \code{sex} column of \code{dt}}

\item{length.unit}{A character argument giving the unit of \code{length}. Will be used in the labels of the figure.}

\item{length.bin.width}{Numeric specifying the increment (delta length) by which length data should be binned to calculate maturity proportions. Use \code{NULL} to remove from the plot.}

\item{split.by.sex}{Logical indicating whether the result should be split by sex.}

\item{filter.exp}{Expression to filter data. E.g. 'sampling_type == "ENS"'}

\item{xlab}{Character giving the x-axis label without unit}

\item{base_size}{Base size parameter for ggplot. See \link[ggplot2]{ggtheme}.}
}
\value{
Returns a ggplot2 or tibble depending on the \code{plot} argument showing the maturity ogives.
}
\description{
Plots an estimate of length or age at 50\% mature for a dataset
}
\details{
Depends on the tidyverse and ggridges packages. The dplyr and ggplot2 packages must be loaded into the workspace.
}
\examples{
# Simple L50 plot
data(survey_ghl)
plot_maturity(survey_ghl, length = "length", maturity = "maturity")
# Split by sex
plot_maturity(survey_ghl, split.by.sex = TRUE)$plot
# A50 plot
plot_maturity(survey_ghl, length = "age", length.unit = "years",
xlab = "Age", length.bin.width = 1, split.by.sex = TRUE)$plot
}
\author{
Mikko Vihtakari // Institute of Marine Research.
}
